requirements.txt
flask
flask-cors
pillow
pydub
python-bidi
